backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
openai==1.3.7
yt-dlp==2023.11.16
ffmpeg-python==0.2.0
boto3==1.34.7
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
numpy==1.26.2
Pillow==10.1.0